requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
bert==2.2.0
boto3==1.17.7
botocore==1.20.7
Brotli==1.0.9
cachetools==4.2.1
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
chart-studio==1.1.0
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
cycler==0.10.0
dash==2.0.0
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
decorator==5.0.6
dill==0.3.3
erlastic==2.0.0
filelock==3.0.12
Flask==2.0.2
Flask-Compress==1.10.1
flatbuffers==1.12
fsspec==2021.10.0
future==0.18.2
gast==0.4.0
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
idna==3.2
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.1
multidict==5.1.0
mysql==0.0.2
mysql-connector-python==8.0.23
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
nlp==0.4.0
nltk==3.5
nodejs==0.1.1
numpy==1.21.2
oauthlib==3.1.0
opt-einsum==3.3.0
optional-django==0.3.0
outcome==1.1.0
packaging==20.9
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
plotly==5.3.1
praw==7.1.4
prawcore==1.5.0
prompt-toolkit==3.0.18
protobuf==3.15.8
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
pymongo==3.11.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.4
pytorch-lightning==1.2.7
pytz==2021.3
PyYAML==5.3.1
react==4.3.0
regex==2020.11.13
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
s3transfer==0.3.4
sacremoses==0.0.44
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
selenium==4.0.0
shrimpy-python==0.0.13
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
tb-nightly==2.5.0a20210411
tenacity==8.0.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0rc0
tensorflow-gpu==2.5.0rc1
tensorflow-hub==0.12.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.10.2
torchmetrics==0.2.0
tqdm==4.56.2
traitlets==5.0.5
transformers==4.5.0
trio==0.19.0
trio-websocket==0.9.2
tweepy==3.10.0
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.7
vaderSentiment==3.3.2
wcwidth==0.2.5
websocket-client==0.57.0
websockets==8.1
Werkzeug==2.0.1
wrapt==1.12.1
wsproto==1.0.0
xxhash==2.0.0
yarl==1.6.3
